66: Add user lens
