+++ app/lenses/meals/job_user_lens.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Meals
+  # Presents a select2 for filtering by a single user.
+  class JobUserLens < ::UserLens
+    protected
+
+    def select2_context
+      "meal_job_lens"
+    end
+  end
+end
